move the settings storage to the user settings frontend
