+++ .changeset/healthy-oranges-fly.md
@@ -1,9 +1,6 @@
 ---
-'@backstage/core-app-api': minor
 '@backstage/plugin-user-settings-backend': minor
 ---
 
-Add new plugin `@backstage/user-settings-backend` to store user related settings
-in the database. Additionally adding a `UserSettingsStorage` implementation of
-the `StorageApi` to easily use the new plugin as drop-in replacement for the
-`WebStorage`.
+Added new plugin `@backstage/plugin-user-settings-backend` to store user related
+settings in the database.
