+++ .changeset/strange-geckos-know.md
@@ -1,0 +1,5 @@
+---
+'@backstage/core-app-api': patch
+---
+
+Clarify that the `WebStorage` observable returns `JsonValue` items.
