+++ packages/core-app-api/package.json
@@ -34,7 +34,6 @@
   "dependencies": {
     "@backstage/config": "^1.0.2-next.0",
     "@backstage/core-plugin-api": "^1.0.6-next.3",
-    "@backstage/errors": "^1.1.1-next.0",
     "@backstage/types": "^1.0.0",
     "@backstage/version-bridge": "^1.0.1",
     "@types/prop-types": "^15.7.3",
