+++ plugins/user-settings/package.json
@@ -32,14 +32,18 @@
     "clean": "backstage-cli package clean"
   },
   "dependencies": {
+    "@backstage/core-app-api": "^1.1.0-next.3",
     "@backstage/core-components": "^0.11.1-next.3",
     "@backstage/core-plugin-api": "^1.0.6-next.3",
+    "@backstage/errors": "^1.1.1-next.0",
     "@backstage/theme": "^0.2.16",
+    "@backstage/types": "^1.0.0",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "4.0.0-alpha.57",
     "@types/react": "^16.13.1 || ^17.0.0",
-    "react-use": "^17.2.4"
+    "react-use": "^17.2.4",
+    "zen-observable": "^0.8.15"
   },
   "peerDependencies": {
     "react": "^16.13.1 || ^17.0.0",
@@ -47,7 +51,6 @@
   },
   "devDependencies": {
     "@backstage/cli": "^0.19.0-next.3",
-    "@backstage/core-app-api": "^1.1.0-next.3",
     "@backstage/dev-utils": "^1.0.6-next.2",
     "@backstage/test-utils": "^1.2.0-next.3",
     "@testing-library/jest-dom": "^5.10.1",
