+++ plugins/user-settings/src/index.ts
@@ -15,11 +15,12 @@
  */
 
 /**
- * A Backstage plugin that provides a settings page
+ * A Backstage plugin that provides various per-user settings functionality.
  *
  * @packageDocumentation
  */
 
+export * from './apis';
 export {
   userSettingsPlugin,
   userSettingsPlugin as plugin,
